requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlmodel>=0.0.21
sqladmin[full]>=0.19
passlib[bcrypt]>=1.7
bcrypt==4.0.1
python-jose[cryptography]>=3.3
python-dotenv>=1.0
httpx>=0.27
aiofiles>=24.1
